+++ app/src/main/java/com/project/drinkly/ui/onboarding/signUp/SignUpNickNameFragment.kt
@@ -9,7 +9,7 @@ import androidx.core.widget.addTextChangedListener
 import com.project.drinkly.R
 import com.project.drinkly.databinding.FragmentSignUpNicknameBinding
 import com.project.drinkly.ui.MainActivity
-import com.project.drinkly.ui.home.HomeMapFragment
+import com.project.drinkly.ui.store.StoreMapFragment
 
 class SignUpNickNameFragment : Fragment() {
 
@@ -41,7 +41,7 @@ class SignUpNickNameFragment : Fragment() {
 
             buttonNext.setOnClickListener {
                 mainActivity.supportFragmentManager.beginTransaction()
-                    .replace(R.id.fragmentContainerView_main, HomeMapFragment())
+                    .replace(R.id.fragmentContainerView_main, StoreMapFragment())
                     .addToBackStack(null)
                     .commit()
             }
